Add files via upload
The index file is the HTML code, the style file is the CSS code and the script is the JAVASRIPTcode. For the background of the first page I used a pic3.png file, for the second I use black.png and for the last page, I used a blackblack.png file.