Add files via upload
index.html: This is likely the main HTML file of your website. It serves as the entry point and defines the structure of your web pages.

style.css: This is the Cascading Style Sheet (CSS) file used to define your web pages' visual appearance and layout. It's used to style the HTML elements in your index.html file.

script.js: This JavaScript file contains code to add interactivity and dynamic behavior to your web pages. You can use JavaScript to respond to user actions, manipulate the DOM (Document Object Model), and perform various tasks on your web pages.

blackblack.png: This image file appears to be used as a background for the last page of your website. Background images are set using CSS.

pic3.png: This image file seems to be used as a background for the first page of your website. Just like the previous one, it's set using CSS.

black.png: This image file is used as a background for the second page of your website. Again, it's set using CSS.